Add DocumentsList component for document display and management
