+++ app/frontend/pages/Document/Document.svelte
@@ -11,36 +11,17 @@
     AlertDialogTrigger,
   } from '$lib/components/ui/alert-dialog'
   import Checkbox from '$lib/components/ui/checkbox/checkbox.svelte'
-  import {
-    DropdownMenu,
-    DropdownMenuContent,
-    DropdownMenuItem,
-    DropdownMenuTrigger,
-  } from '$lib/components/ui/dropdown-menu'
   import { Separator } from '$lib/components/ui/separator'
   import { Link, router } from '@inertiajs/svelte'
   import { ListChecks } from '@lucide/svelte'
-  import {
-    DotsVertical,
-    File06 as File,
-    Trash01,
-    X,
-  } from '@voolt_technologies/untitledui-svelte'
-  import { formatDistanceToNowStrict } from 'date-fns'
-  import { ptBR } from 'date-fns/locale'
+  import { Trash01, X } from '@voolt_technologies/untitledui-svelte'
   import type { DocumentType } from './types'
-
+  import DocumentsList from '$lib/components/documents/list.svelte'
   type Props = {
     documents: DocumentType[]
   }
 
   let { documents }: Props = $props()
-
-  const daysAgo = (date: Date) =>
-    formatDistanceToNowStrict(date, {
-      locale: ptBR,
-      roundingMethod: 'ceil',
-    })
 
   let editMode = $state(false)
   let selectedDocuments: number[] = $state([])
@@ -56,29 +37,10 @@
     return selectedDocuments.length === documents.length
   }
 
-  function toggleSelect(isChecked: boolean, id: number) {
-    if (isChecked) {
-      selectedDocuments.push(id)
-    } else {
-      selectedDocuments = selectedDocuments.filter(
-        (documentId) => documentId !== id
-      )
-    }
-
-    return selectedDocuments.includes(id)
-  }
-
   function resetEditMode() {
     editMode = false
     selectedDocuments = []
     openBulkDelete = false
-  }
-
-  function removeDocument(hashId: string) {
-    router.delete(`/documents/${hashId}`, {
-      preserveState: true,
-      preserveScroll: true,
-    })
   }
 
   function destroyAll() {
@@ -136,81 +98,21 @@
 
 <div class="mt-2 space-y-1">
   {#each documents as document (document.id)}
-    <Link
-      prefetch
-      href={`/documents/${document.sqid}`}
-      class="grid grid-cols-[1fr_auto_auto] items-center hover:bg-muted py-2 px-4 rounded cursor-pointer"
-    >
-      <div class="flex items-center gap-4">
-        {#if editMode}
-          <Checkbox
-            bind:checked={
-              () => selectedDocuments.includes(document.id),
-              (isChecked) => toggleSelect(isChecked, document.id)
-            }
-          />
-        {/if}
-        <File class="size-5 text-muted-foreground" />
-        <div class="font-brand flex flex-col font-semibold max-w-60">
-          <span class="text-accent-foreground text-sm">
-            {document.title || 'Sem título'}
-          </span>
-
-          <span class="text-muted-foreground text-xs font-medium truncate">
-            {document.description || 'Lorem ipsum solar sit amet...'}
-          </span>
-        </div>
-      </div>
-
+    {#if editMode}
       <div
-        class="grid grid-cols-[100px_100px_3.5rem] text-sm font-brand font-medium gap-4 pr-2 text-right"
+        class="grid grid-cols-[1fr_auto_auto] items-center hover:bg-muted py-2 px-4 rounded cursor-pointer"
       >
-        <span class="text-accent-foreground tabular-nums">
-          {daysAgo(new Date(document.created_at))}
-        </span>
-        <span class="text-accent-foreground">
-          {daysAgo(new Date(document.updated_at))}
-        </span>
+        <DocumentsList {document} {editMode} bind:selectedDocuments />
       </div>
-
-      <AlertDialog>
-        <DropdownMenu>
-          <DropdownMenuTrigger
-            class="cursor-pointer p-1 hover:bg-sidebar-accent rounded-md"
-          >
-            <DotsVertical class="size-4 text-accent-foreground" />
-          </DropdownMenuTrigger>
-
-          <DropdownMenuContent class="mr-2">
-            <AlertDialogTrigger>
-              <DropdownMenuItem
-                class="text-destructive data-[highlighted]:bg-destructive/10 data-[highlighted]:text-destructive"
-              >
-                <Trash01 />
-                <span class="font-medium font-brand"> Mover para lixeira </span>
-              </DropdownMenuItem>
-            </AlertDialogTrigger>
-          </DropdownMenuContent>
-        </DropdownMenu>
-        <AlertDialogContent>
-          <AlertDialogHeader>
-            <AlertDialogTitle>Tem certeza disso?</AlertDialogTitle>
-            <AlertDialogDescription>
-              Você está prestes a mover o documento <strong
-                >{document.title}</strong
-              > para a lixeira. Isso significa que você não poderá acessá-lo até
-              que o restaure. Você tem certeza de que deseja continuar?
-            </AlertDialogDescription>
-          </AlertDialogHeader>
-          <AlertDialogFooter class="gap-4">
-            <AlertDialogCancel>Cancelar</AlertDialogCancel>
-            <AlertDialogRemove onclick={() => removeDocument(document.sqid)}
-              >Remover</AlertDialogRemove
-            >
-          </AlertDialogFooter>
-        </AlertDialogContent>
-      </AlertDialog>
-    </Link>
+    {:else}
+      <Link
+        prefetch
+        href={`/documents/${document.sqid}`}
+        class="grid grid-cols-[1fr_auto_auto] items-center hover:bg-muted py-2 px-4 rounded cursor-pointer"
+      >
+        <DocumentsList {document} {editMode} bind:selectedDocuments />
+      </Link>
+    {/if}
   {/each}
 </div>
